server/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.9
gunicorn==20.0.4
ibm-cloud-sdk-core==3.10.0
ibmcloudant==0.0.34
idna==3.2
jproperties==2.1.1
Pillow==8.0.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
ratelimit==2.2.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.0
urllib3==1.26.6
websocket-client==1.1.0
